[ADF-968]- moved site service into core
