+++ ng2-components/ng2-activiti-analytics/src/models/index.ts
@@ -17,3 +17,8 @@
 
 export * from './chart.model';
 export * from './report.model';
+
+export * from './card-view-textitem.model';
+export * from './card-view-dateitem.model';
+export * from './file.model';
+export * from './site.model';
