+++ ng2-components/ng2-alfresco-core/src/components/view/card-view-textitem.component.spec.ts
@@ -51,7 +51,7 @@ describe('CardViewTextItemComponent', () => {
     beforeEach(() => {
         fixture = TestBed.createComponent(CardViewTextItemComponent);
         component = fixture.componentInstance;
-        component.property = <CardViewTextItemModel>{
+        component.property = <CardViewTextItemModel> {
             type: 'text',
             label: 'Text label',
             value: 'Lorem ipsum',
