+++ ng2-components/ng2-alfresco-core/src/services/thumbnail.service.spec.ts
@@ -19,10 +19,10 @@ import { async, TestBed } from '@angular/core/testing';
 import { HttpModule } from '@angular/http';
 import { CookieServiceMock } from './../assets/cookie.service.mock';
 import { AlfrescoApiService } from './alfresco-api.service';
-import { AuthenticationService } from './authentication.service';
 import { AlfrescoContentService } from './alfresco-content.service';
 import { AlfrescoSettingsService } from './alfresco-settings.service';
 import { AppConfigService } from './app-config.service';
+import { AuthenticationService } from './authentication.service';
 import { CookieService } from './cookie.service';
 import { LogService } from './log.service';
 import { StorageService } from './storage.service';
