+++ ng2-components/ng2-alfresco-documentlist/README.md
@@ -18,6 +18,9 @@
 - [Breadcrumb Component](#breadcrumb-component)
   * [Properties](#properties-1)
   * [Events](#events-1)
+- [Dropdown Site Component](#dropdown-site-component)
+  * [Properties](#properties-2)
+  * [Events](#events-2)
 - [Menu Actions](#menu-actions)
 - [Custom columns](#custom-columns)
   * [DataColumn Properties](#datacolumn-properties)
@@ -33,7 +36,7 @@
     + [Folder actions](#folder-actions)
   * [Context Menu](#context-menu)
   * [Navigation mode](#navigation-mode)
-  * [Events](#events-2)
+  * [Events](#events-3)
 - [Advanced usage and customization](#advanced-usage-and-customization)
   * [Custom row filter](#custom-row-filter)
   * [Custom image resolver](#custom-image-resolver)
